before merge; also progress on navbar styling
