+++ client/components/navbar/search/ChooseDept.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+function ChooseDept(){
+  return(
+    <button className="chooseDepartment">All Departments</button>
+  )
+}
+export default ChooseDept
